Preparación Servicio de música sin funcionalidad en DetalleRutas. Cambio de int a double fallo distancia

+++ SendeGo/app/src/main/java/com/app/sendego/activities/RouteRegistration.java
@@ -93,9 +93,9 @@ public class RouteRegistration extends AppCompatActivity {
                 return;
             }
 
-            int distance;
+            double distance;
             try {
-                distance = Integer.parseInt(distanceStr);
+                distance = Double.parseDouble(distanceStr);
             } catch (NumberFormatException e) {
                 Toast.makeText(this, "La distancia debe ser un número válido", Toast.LENGTH_SHORT).show();
                 return;
